Refactor: data 모듈 패키지 구조 변경
- MemberPersistentAdapter를 Adapter 모듈로 이동
  - JAP 의존성을 api로 변경
